requirements.txt
Flask
paho-mqtt==1.6.1



